requirements.txt
joblib==1.3.2
scikit-learn==1.3.2
numpy==1.26.0
pillow==10.0.0